# [PERF] Jackson reflection-free serialization/deserialization