--- a/docs/guides/server/configuration-production.adoc
+++ b/docs/guides/server/configuration-production.adoc
@@ -1,6 +1,7 @@
 <#import "/templates/guide.adoc" as tmpl>
 <#import "/templates/kc.adoc" as kc>
 <#import "/templates/links.adoc" as links>
+<#import "/templates/features.adoc" as features>
 
 <@tmpl.guide
 title="Configuring {project_name} for production"
@@ -88,13 +89,13 @@ export JAVA_OPTS_APPEND="-Djava.net.preferIPv4Stack=false -Djava.net.preferIPv6A
 See <@links.server id="caching" anchor="network-bind-address"/> for more details.
 
 == Better HTTP performance
-NOTE: This section is describing a feature which is currently in *preview* as we gather more feedback about potential issues. We appreciate any feedback.
+<@features.techpreview feature="http-optimized-serializers" additionalCommunityText="We gather more feedback on this feature to promote it to supported. Please, share your feedback about any issue in https://github.com/keycloak/keycloak/discussions/43484[this discussion]."/>
 
 In production environments, the performance of the HTTP layer is critical.
 Every request passes through it, making it a key factor in overall system responsiveness, scalability, and user experience.
 
-This feature introduces Jackson JSON reflection-free serializers to optimize HTTP processing.
-By removing the overhead of reflection during serialization/deserialization, it delivers measurable benefits:
+This feature improves how {project_name} handles JSON data in HTTP requests and responses.
+The result is a more efficient runtime with measurable benefits:
 
 - ~5% increase in throughput
 - More stable response times
@@ -107,7 +108,5 @@ The only known tradeoff is that build time increases by ~6% as certain actions w
 You can enable this feature as follows:
 
 <@kc.start parameters="--features=http-optimized-serializers"/>
-
-The feature is *disabled by default*, as we gather more feedback.
 
 </@tmpl.guide>