- Closes #42945

I've followed up on the [guide](https://quarkus.io/blog/quarkus-metaprogramming/) provided by the Quarkus team. The Jackson-free serdes was mentioned during the presentation about Observability performance in Quarkus, so they're including that even in their benchmarks.

In the guide, there's mentioned ~12% improvement in the throughput for HTTP calls. I made a brief performance analysis, and these are my findings:

## Performance analysis

- Run official KC distribution (OLD) and official KC distribution with the optimization (NEW)
- Used `wrk` from the [Hyperfoil](https://github.com/Hyperfoil/Hyperfoil)

### Test characteristic
- Benchmarking how many req/s can KC server handle on one node
- Only the `host:8080/admin/realms/master/clients` endpoint was called (multiple clients that are deserialized) with the access token 
- On a client, changed how many threads and connections are used when introducing the load on the server
-- Matrix 1: 3 threads, 3 connections
-- Matrix 2: 10 threads, 10 connections
-- Matrix 3: 15 threads, 15 connections
- Every matrix run for 15s
- Every matrix is executed 5x

### Matrix 1 (3 threads, 3 connections)
**Requests per second**
| Run | Old  | New  | Diff % |
|-----|-----------|-----------|-------------|
| 1   | 1224      | 1385      | ⬆️ **13.15** ✅ |
| 2   | 1494      | 1486      | ⬇️ *-0.54* ❌ |
| 3   | 1483      | 1550      | ⬆️ **4.52** ✅ |
| 4   | 1501      | 1491      | ⬇️ *-0.67* ❌ |
| 5   | 1492      | 1634      | ⬆️ **9.52** ✅ |

Average: **5.196% improvement** ✅

### Matrix 2 (10 threads, 10 connections)
**Requests per second**
| Run | Old  | New  | Diff % |
|-----|-----------|-----------|-------------|
| 1   | 3455      | 3636      | ⬆️ **5.24** ✅ |
| 2   | 3416      | 3504      | ⬆️ **2.58** ✅ |
| 3   | 3395      | 3496      | ⬆️ **2.97** ✅ |
| 4   | 3341      | 3419      | ⬆️ **2.33** ✅ |
| 5   | 3347      | 3738      | ⬆️ **11.68** ✅ |

Average: **4.96% improvement** ✅

### Matrix 3 (15 threads, 15 connections)
**Requests per second**
| Run | Old  | New  | Diff % |
|-----|-----------|-----------|-------------|
| 1   | 3268      | 3100      | ⬇️ *-5.14* ❌ |
| 2   | 4034      | 4278      | ⬆️ **6.05** ✅ |
| 3   | 3976      | 4586      | ⬆️ **15.34** ✅ |
| 4   | 4000      | 4272      | ⬆️ **6.80** ✅ |
| 5   | 3993      | 4281      | ⬆️ **7.21** ✅ |

Average: **6.052% improvement** ✅

### Augmentation time
As some work is moved to the build phase, we experience a small increase in the augmentation time (which should be ok):

| Run | Old  | New  | Diff | Diff % |
|-----|------|------|------|--------|
| 1   | 4341ms | 4501ms | ⬆️ *160ms* ❌ | ⬆️ *3.69%* ❌ |
| 2   | 4482ms | 5131ms | ⬆️ *649ms* ❌ | ⬆️ *14.48%* ❌ |
| 3   | 8233ms | 8630ms | ⬆️ *397ms* ❌ | ⬆️ *4.82%* ❌ |
| 4   | 7552ms | 7691ms | ⬆️ *139ms* ❌ | ⬆️ *1.84%* ❌ |

Average Old: **6152ms**
Average New: **6488ms**
Difference: **+336ms** ❌
Difference in %: **6.20%** ❌

### Node characteristics
```
Architecture:                x86_64
  CPU op-mode(s):            32-bit, 64-bit
  Address sizes:             46 bits physical, 48 bits virtual
  Byte Order:                Little Endian
CPU(s):                      20
  On-line CPU(s) list:       0-19
Vendor ID:                   GenuineIntel
  Model name:                12th Gen Intel(R) Core(TM) i7-1280P
    CPU family:              6
    Model:                   154
    Thread(s) per core:      2
    Core(s) per socket:      14
Memory:                      32GB
```

## Conclusion

In total, in this brief performance analysis, we experience **5.40%**✅ improvement in the throughput in our API. Note that values for some runs differ as it might be influenced by other tasks done on the CPU and not related to this benchmark. This test only gets clients from the realm, so for some endpoints, where the serdes is not used much does not have a big effect. On the other hand, the effect should be higher when serdes is used frequently. 

I think it's a nice improvement, and there should not be any issues I'm not aware of. 

cc: @keycloak/cloud-native @keycloak/sre
cc: @franz1981 @brunobat