@pruivo 2025-09-25
@mabartos, what is happening in Matrix 1? Win/Lose alternates between runs.
How many clients and what is the JDK version?