@mabartos 2025-09-25
> what is happening in Matrix 1? Win/Lose alternates between runs.

We can't ensure the runs are 1:1 mapped, f.e. there might be the case when some run for the New loses,  but when compared with other run, it'd win. You can see that for the matrix 1, these values oscillates for the Old in range 1224-1492, where for the New 1385-1634. I haven't checked stdev and other props. 

> How many clients and what is the JDK version?

The default ones - ~5? OpenJDK 21.0.3

@pruivo It'd be nice to compare it in the KC benchmark that could show more interesting data. I can maybe also provide a comparison of flame graphs.