@mabartos 2025-09-25
> And in term of native version, it is faster? Or no difference?

@franz1981 We don't use native in Keycloak. Or are you referring to sth else?