@franz1981 2025-09-25
I remember an issue where keycloak got bad XP with native image but I wasn't aware is not available as native app 🙏