@franz1981 2025-09-25
Just FYI, loolking how reflection work, it got bad and worse as the number and type of fields grows and the number of serialized/deserialized types too. Not to mention the RSS consumed due to method handle generation, which won't happen in reflection free case