@pruivo 2025-09-25
@mabartos I run the benchmark locally, so take the results with a grain of salt as I was using the PC while benchmarking. I run twice, and your PR seems more stable. The committed memory at the end of the test is lower with your PR. 

### This PR
```
[benchamrk-run] ================================================================================
[benchamrk-run] ---- Global Information --------------------------------------------------------
[benchamrk-run] > request count                                     793000 (OK=793000 KO=0     )
[benchamrk-run] > min response time                                      2 (OK=2      KO=-     )
[benchamrk-run] > max response time                                    869 (OK=869    KO=-     )
[benchamrk-run] > mean response time                                    16 (OK=16     KO=-     )
[benchamrk-run] > std deviation                                         21 (OK=21     KO=-     )
[benchamrk-run] > response time 50th percentile                          9 (OK=9      KO=-     )
[benchamrk-run] > response time 75th percentile                         16 (OK=16     KO=-     )
[benchamrk-run] > response time 95th percentile                         62 (OK=62     KO=-     )
[benchamrk-run] > response time 99th percentile                        102 (OK=102    KO=-     )
[benchamrk-run] > mean requests/sec                                1099.86 (OK=1099.86 KO=-     )
[benchamrk-run] ---- Response Time Distribution ------------------------------------------------
[benchamrk-run] > t < 800 ms                                        792999 (   100%)
[benchamrk-run] > 800 ms <= t < 1200 ms                                  1 (     0%)
[benchamrk-run] > t >= 1200 ms                                           0 (     0%)
[benchamrk-run] > failed                                                 0 (     0%)
[benchamrk-run] ================================================================================
```
### Main (commit 29bee21683ad987e2a3c313a5caebca7f0063a9e)
```
[benchamrk-run] ================================================================================
[benchamrk-run] ---- Global Information --------------------------------------------------------
[benchamrk-run] > request count                                     793000 (OK=793000 KO=0     )
[benchamrk-run] > min response time                                      2 (OK=2      KO=-     )
[benchamrk-run] > max response time                                   2823 (OK=2823   KO=-     )
[benchamrk-run] > mean response time                                    54 (OK=54     KO=-     )
[benchamrk-run] > std deviation                                        213 (OK=213    KO=-     )
[benchamrk-run] > response time 50th percentile                         10 (OK=10     KO=-     )
[benchamrk-run] > response time 75th percentile                         21 (OK=21     KO=-     )
[benchamrk-run] > response time 95th percentile                        153 (OK=153    KO=-     )
[benchamrk-run] > response time 99th percentile                       1246 (OK=1242   KO=-     )
[benchamrk-run] > mean requests/sec                                1099.86 (OK=1099.86 KO=-     )
[benchamrk-run] ---- Response Time Distribution ------------------------------------------------
[benchamrk-run] > t < 800 ms                                        779570 ( 98.31%)
[benchamrk-run] > 800 ms <= t < 1200 ms                               4380 (  0.55%)
[benchamrk-run] > t >= 1200 ms                                        9050 (  1.14%)
[benchamrk-run] > failed                                                 0 (     0%)
[benchamrk-run] ================================================================================
```