@mabartos 2025-09-26
@pruivo Great! I think we have enough data to proceed with the approach. 

@keycloak/cloud-native Could anyone look at this? Thanks!