@brunobat 2025-09-26
Did you measured the startup time?