@ahus1 2025-09-26
I would be +1 to use it, at the same time Quarkus is stating this is a [technology preview feature](https://quarkus.io/version/3.27/guides/rest#reflection-free-jackson-serialization-and-deserialization). 

So I'd say we would need to wrap it as an opt-in preview feature in KC as we don't want to support it downstream, and some tests enabled and disabled. We might open a ticket with the Quarkus team make it non-preview eventually. 

Reading the docs, the following part worries me a bit: 

> Developers can further customize JSON processing by implementing the ObjectMapperCustomizer interface. This interface allows fine-grained control over the ObjectMapper, enabling the registration of custom serializers and deserializers while ensuring compatibility with the reflection-free optimization. If additional configuration is needed, implement an ObjectMapperCustomizer bean and register any necessary modules or settings.