@ahus1 2025-09-26
@brunobat / @franz1981 - thank for joining this discussion. Do you have additional information for us on the maturity of the feature and any plans for graduation of this feature?