@brunobat 2025-09-26
I believe @franz1981 has details about why it's still in preview.