@brunobat 2025-09-26
> > Did you measured the startup time?
> 
> @brunobat Nope. Should we expect some bigger differences?
> 
> > I would be +1 to use it, at the same time Quarkus is stating this is a [technology preview feature](https://quarkus.io/version/3.27/guides/rest#reflection-free-jackson-serialization-and-deserialization).
> 
> @ahus1 Oh, totally missed this. Thanks for pointing this out! Yes, unfortunately, we'd need to have it disabled by default :(

The expectation is that it will start faster because more work was done at build, however, I think we should measure the real world behavior differences in the C2 compiler under load.