@mabartos 2025-09-26
> Maybe you could provide the option with a default value to false for a while and then switch it to true at some point (and keeping the ability to disable it?).

Yes, that's the plan. 

Thank you all for your feedback!